+++ eyeball/src/shared.rs
@@ -137,8 +137,8 @@ impl<T> Observable<T> {
     /// Update the inner value and notify subscribers.
     ///
     /// Note that even if the inner value is not actually changed by the
-    /// closure, subscribers will be notified as if it was. Use one of the
-    /// other update methods below if you want to conditionally mutate the
+    /// closure, subscribers will be notified as if it was. Use
+    /// [`update_if`][Self::update_if] if you want to conditionally mutate the
     /// inner value.
     pub fn update(&self, f: impl FnOnce(&mut T)) {
         self.state.write().unwrap().update(f);
@@ -150,24 +150,6 @@ impl<T> Observable<T> {
     /// should be notified of a change to the inner value.
     pub fn update_if(&self, f: impl FnOnce(&mut T) -> bool) {
         self.state.write().unwrap().update_if(f);
-    }
-
-    /// Update the inner value and notify subscribers if the updated value does
-    /// not equal the previous value.
-    pub fn update_eq(&self, f: impl FnOnce(&mut T))
-    where
-        T: Clone + PartialEq,
-    {
-        self.state.write().unwrap().update_eq(f);
-    }
-
-    /// Update the inner value and notify subscribers if the hash of the updated
-    /// value does not equal the hash of the previous value.
-    pub fn update_hash(&self, f: impl FnOnce(&mut T))
-    where
-        T: Hash,
-    {
-        self.state.write().unwrap().update_hash(f);
     }
 
     /// Get the number of references to the inner value.
@@ -264,8 +246,8 @@ impl<'a, T> ObservableWriteGuard<'a, T> {
     /// Update the inner value and notify subscribers.
     ///
     /// Note that even if the inner value is not actually changed by the
-    /// closure, subscribers will be notified as if it was. Use one of the
-    /// other update methods below if you want to conditionally mutate the
+    /// closure, subscribers will be notified as if it was. Use
+    /// [`update_if`][Self::update_if] if you want to conditionally mutate the
     /// inner value.
     pub fn update(this: &mut Self, f: impl FnOnce(&mut T)) {
         this.inner.update(f);
@@ -277,24 +259,6 @@ impl<'a, T> ObservableWriteGuard<'a, T> {
     /// should be notified of a change to the inner value.
     pub fn update_if(this: &mut Self, f: impl FnOnce(&mut T) -> bool) {
         this.inner.update_if(f);
-    }
-
-    /// Update the inner value and notify subscribers if the updated value does
-    /// not equal the previous value.
-    pub fn update_eq(this: &mut Self, f: impl FnOnce(&mut T))
-    where
-        T: Clone + PartialEq,
-    {
-        this.inner.update_eq(f);
-    }
-
-    /// Update the inner value and notify subscribers if the hash of the updated
-    /// value does not equal the hash of the previous value.
-    pub fn update_hash(this: &mut Self, f: impl FnOnce(&mut T))
-    where
-        T: Hash,
-    {
-        this.inner.update_hash(f);
     }
 }