+++ eyeball/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Remove `T: Clone` bound from `set_eq`
 - Merge `replace`s functionality of returning the previous value into `set`
 - Return the previous inner value if `set_eq` or `set_hash` replaces it
+- Add `update_if`
 
 # 0.4.2